Mantra.Chain.Internal

Description

Extracting scripts from witnesses.

Scripts

interpretAsScript

Arguments

:: Timelock StandardCrypto
The witness.

-> Maybe (SimpleScript SimpleScriptV2, ScriptHash)
The script and its hash, if the witness was a script.

Convert a witness to a script.

toScript

Arguments

:: SimpleScript SimpleScriptV2
The simple script.

-> (SimpleScript SimpleScriptV2, ScriptHash)
The script and its hash.

Convert a simple script to a script and its hash.

toSimpleScriptV2

Arguments

:: String
The string representation.

-> Maybe (SimpleScript SimpleScriptV2)
The simple script, if the parsing succeeded.

Parse a string representation of a simple script.